+++ src/users/dto/user-summary.dto.ts
@@ -4,5 +4,6 @@ export class UserSummaryDto {
   id: number;
   firstName?: string;
   lastName?: string;
+  email?: string;
   profileFile?: Partial<FileEntity> | null;
 }